requirements.txt
alembic==1.7.5
aniso8601==9.0.1
APScheduler==3.8.1
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
colorhash==1.0.4
configparser==5.2.0
cryptography==36.0.1
ecdsa==0.17.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-MonitoringDashboard==3.1.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.2
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
numpy==1.21.5
packaging==21.3
parameterized==0.8.1
platformdirs==2.4.0
pluggy==1.0.0
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
PyMySQL==1.0.2
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
rsa==4.8
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.28
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==2.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
